feat(delete): move to tools processing dir
