+++ geoplateforme/processing/tools/delete_stored_data.py
